+++ TP_Quest/src/main/java/application/MessagingStompWebsocketApplication.java
@@ -1,4 +1,4 @@
-package server.websocketTest;
+package application;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
