+++ TP_Quest/src/main/java/application/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package server.websocketTest;
+package application;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
