+++ TP_Quest/src/main/java/controller/QuestionController.java
@@ -1,25 +1,24 @@
-package server.websocketTest;
+package controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import service.*;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.stereotype.Controller;
-import dto.QuestionDto;
-import dto.QuestionQueryDto;
+import dto.*;
+import service.impl.QuestionServiceImpl;
 
 @Controller
 @RequiredArgsConstructor
 @Slf4j
 public class QuestionController {
 
-    private final QuestionService instance;
+    private final QuestionServiceImpl instance;
 
     @MessageMapping("/question")
     @SendTo("/topic/questions")
     public QuestionDto randomQuestionDto(QuestionQueryDto questionDtoQuery) {
+        System.out.println(instance.toString());
         return instance.getRandomQuestion(
                 questionDtoQuery.getTopic(),
                 questionDtoQuery.getDifficulty());
