+++ TP_Quest/src/main/java/dao/domain/Question.java
@@ -4,9 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
