+++ TP_Quest/src/main/java/service/QuestionService.java
@@ -1,37 +1,13 @@
 package service;
 
-import dao.domain.Question;
-import dao.QuestionDao;
 import dto.QuestionDto;
-import lombok.RequiredArgsConstructor;
-import mapper.QuestionMapper;
-import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class QuestionService {
+public interface QuestionService {
+    List<QuestionDto> getQuestions(long id_topic, long id_difficulty);
 
-    private final QuestionDao questionDAO;
+    String getCorrectAnswer(long id);
 
-    public List<QuestionDto> getQuestions(long id_topic, long id_difficulty) {
-        List<QuestionDto> questions = new ArrayList<>();
-        List<Question> questionsDao = questionDAO.getQuestions(id_topic, id_difficulty);
-        for (var user : questionsDao) {
-            questions.add(QuestionMapper.questionToDto(user));
-        }
-        return questions;
-    }
-
-    public String getCorrectAnswer(long id) {
-        return questionDAO.getCorrectAnswer(id);
-    }
-
-    public QuestionDto getRandomQuestion(String topic, String difficulty) {
-        return QuestionMapper.questionToDto(questionDAO.getRandomQuestion(topic, difficulty));
-    }
-
+    QuestionDto getRandomQuestion(String topic, String difficulty);
 }
-
