+++ TP_Quest/src/main/resources/application.yaml
@@ -7,4 +7,3 @@ spring:
 
   flyway:
     baseline-on-migrate: true
-
