+++ TP_Quest/src/main/resources/db/migration/V1__Create_Tables.sql
@@ -9,16 +9,16 @@ VALUES (100),(200),(300),(400),(500);
 
 CREATE TABLE IF NOT EXISTS Topic(
 			id SERIAL PRIMARY KEY,
-            name VARCHAR(50) NOT NULL,
+            name VARCHAR(100) NOT NULL,
   			UNIQUE(name)
 );
 INSERT INTO Topic(name)
-VALUES ('В мире животных'),('Культура и традиции'),('Космос'),('История'),('Литература'),('Страны мира'),('Кино');
+VALUES ('Космос'),('Культура и традиции'),('Игры'),('Великие деятели'),('Страны мира'),('Литература'),('Кино');
 
 CREATE TABLE IF NOT EXISTS Questions(
                        id SERIAL PRIMARY KEY,
-                       text VARCHAR(150) NOT NULL,
-  					   answer VARCHAR(50) NOT NULL,
+                       text VARCHAR(300) NOT NULL,
+  					   answer VARCHAR(150) NOT NULL,
                        id_topic INTEGER REFERENCES topic(id),
   					   id_difficulty INTEGER REFERENCES difficulty(id),
   					   UNIQUE(text)
