+++ TP_Quest/src/main/resources/db/migration/V3__Insert_Values.sql
@@ -11,8 +11,7 @@ VALUES  ('ЕГО поднял над Скалой Прайда мандрил Р
         ('ЭТОТ ФИЛЬМ далек от литературного первоисточника. Например, Гудвин там становится убийцей','Короли и капуста',7,4),
         ('В фильме 1926 года ОН становится командиром подразделения Рабоче-крестьянской Красной Армии','Беня Крик',7,5),
         ('Крокодил Гена пел: Прилетит вдруг волшебник В голубом вертолете, И бесплатно покажет кино. Автор музыки к этой песне..','Владимир Шаинский',7,3),
-        ('Этот японский поэт Х в. составил первую антологию японской поэзии (достаточно фамилии),',7,7,2),
         ('Так назывался первый в истории группы Кино альбом, записанный в профессиональной студии','Звезда по имени Солнце',7,1),
         ('В пионерской песне с рефреном неправильно, неправильно кончается кино! осуждается его страшная гибель.','Чапаева ',7,4),
-        (' Именно он был первым ведущим передачи Кинопанорама','Алексей Каплер.',7,5),
-        (' ТАК называется биографический фильм о жизни рок-музыканта Иэна Дьюри','Секс, наркотики и рок-н-ролл',7,4);
+        ('Именно он был первым ведущим передачи Кинопанорама','Алексей Каплер.',7,5),
+        ('ТАК называется биографический фильм о жизни рок-музыканта Иэна Дьюри','Секс, наркотики и рок-н-ролл',7,4);
