+++ TP_Quest/src/main/resources/db/migration/V4_1__TestData.sql
@@ -1,0 +1,46 @@
+INSERT INTO Difficulty(score)
+VALUES (600),(700),(800),(900),(1000);
+
+INSERT INTO questions(text,answer,id_topic,id_difficulty)
+VALUES ('Какая по счету от Солнца планета Земля?', 'Третья', 1, 1),
+       ('Плутон является планетой? (да/нет)', 'Нет', 1, 2),
+       ('Самая крупная звезда в Солнечной системе?', 'Солнце', 1, 3),
+       ('Сколько всего планет в Солнечной системе?', 'Восемь', 1, 4),
+       ('Какая ближайшая к Солнцу планета?', 'Меркурий', 1, 5),
+       ('Самая большая планета Солнечной системы?', 'Юпитер', 1, 6),
+       ('Эта планета могла стать звездой, но не набрала достаточно массы.', 'Юпитер', 1, 7),
+       ('Кто был первооткрывателем законов движения планет Солнечной системы? (Имя, Фамилия)', 'Иоганн Кеплер', 1, 8),
+       ('Сколько спутников у Марса?', 'Два', 1, 9),
+       ('Какой самый большой спутник в Солнечной системе?', 'Ганимед', 1, 10);
+
+INSERT INTO game_user(name, password)
+VALUES ('Tirry', '12345'),
+       ('Bungo', '9021pu'),
+       ('Кирилл', 'liachovi'),
+       ('BigO', '12345'),
+       ('Злата', 'tr145rt'),
+       ('liz-ka', '100ka'),
+       ('grapestacy', 'greengrs'),
+       ('Stacy', '123456'),
+       ('Nothing_else', 'qwerty'),
+       ('Nothing0else', 'hornsan'),
+       ('Biggy', 'bongbong'),
+       ('yummy', '098765'),
+       ('liz', 'ochka'),
+       ('Tiger', '012356'),
+       ('SomeUser', 'gfhju'),
+       ('JustUser', 'asdfgh'),
+       ('NotUser', 'zxcvbn'),
+       ('Zlatik', 'mnbvc'),
+       ('Ooops', 'zpftfkcz'),
+       ('Zlattik', 'hoho2021'),
+       ('Subs', 'gfvfubnb');
+
+
+
+
+
+
+
+
+
