+++ TP_Quest/src/main/resources/db/migration/V4_2__TestData.sql
@@ -1,0 +1,64 @@
+INSERT INTO questions(text,answer,id_topic,id_difficulty)
+VALUES
+    ('Хачапури – сочная тягучая лепешка с начинкой из сыра или яйца. Из какой страны родом это блюдо?','Грузия',2,1),
+    ('Вареники - это изделия из пресного теста с начинкой из рубленого мяса, овощей, грибов, фруктов, творога или ягод. Из какой страны родом это блюдо? ','Украина',2,2),
+    ('Пути́н - это очень калорийное блюдо, состоящее из картофеля фри, сыра и специально приготовленного соуса. Из какой страны оно родом?','Канада',2,3),
+    ('Шаурма – это одно из самых доступных для любого кошелька блюд во всем мире. В какой стране началась история этого кулинарного хита? ','ОАЭ',2,4),
+    ('Панчхан - это блюдо состоит из множества закусок и салатов, которые подаются на небольших тарелках как аккомпанемент к основному блюду и рису. Из какой страны оно родом?','Южная Корея',2,5),
+    ('Брынзовые галушки - это маленькие картофельные клецки, заправленные мягким сыром и беконом. Из какой страны родом это блюдо?','Словакия',2,6),
+    ('Адобо - это блюдо, состоящее из кусочков курицы или свинины, тушеных в смеси соевого соуса и уксуса и сдобренных перцем, чесноком и лавровым листом. Из какой страны оно родом?','Филиппины',2,7),
+    ('Кацудон – это тонкая свиная отбивная с яйцом, которую кладут сверху в чашу с рисом. Из какой страны родом это блюдо?','Япония',2,8),
+    ('Билтонг - разновидность вяленого мяса. Чаще всего билтонг готовят из говядины, но он прекрасно получается и из других видов мяса, например, страусятины. Из какой страны родом это блюдо?','ЮАР',2,9),
+    ('Пирог-поплавок - мясной пирог, плавающий в тарелке с густым гороховым супом. Из какой страны родом это блюдо? ','Австралия',2,10);
+
+INSERT INTO questions(text,answer,id_topic,id_difficulty)
+VALUES
+    ('Сколько костяшек входит в стандартный набор в классическом домино?','Двадцать восемь',3,1),
+    ('В какой стране (кроме Индии) появилось Домино?)','Китай',3,2),
+    ('Сколько человек играет в классическом Домино?(2-6, 2-4, 4-6)','2-4',3,3),
+    ('Если играют 2 человека, то сдают по ... костяшек.','Семь',3,4),
+    ('Если играют 3 или 4 человека, то сдают по ... костяшек.','Пять',3,5),
+    ('Первый ход принадлежит игроку, у которого любой минимальный дубль кроме','0:0',3,6),
+    ('Как называются костяшки, которые остались после раздачи игрокам и помещаются в резерв, который в классическом домино?','Базар',3,7),
+    ('В каком веке в Китае впервые достоверно упоминается домино?','Тринадцатом',3,8),
+    ('Если при подсчете очков у частника осталась всего одна костяшка с дублем 0:0, то она засчитывается, как ... очков.','Двадцать пять',3,9),
+    ('Как называется ситуация в домино, когда у игроков есть костяшки на руках, а ходить нечем?','Рыба',3,10);
+
+INSERT INTO questions(text,answer,id_topic,id_difficulty)
+VALUES
+    ('Какой русский ученый открыл периодический закон химических элементов - фундаментальный закон естествознания в России/СССР? (Фамилия)','Менделеев',4,1),
+    ('Какой русский ученый впервые описал вирусный гепатит А в России/СССР? (Фамилия)','Боткин',4,2),
+    ('Какой русский ученый создал первый фотоэлемент (прибор, преобразующий энергию фотонов в электричество) в России/СССР? (Фамилия)','Столетов',4,3),
+    ('Какая фамилия авиаконструктора, который разработал первый в мире пассажирский сверхзвуковой авиалайнер в России/СССР?','Туполев',4,4),
+    ('Какая фамилия нобелевского лауреата, создавшего науку о высшей нервной деятельности?','Павлов',4,5),
+    ('Кто в 2002 году доказал теорему Пуанкаре (одну из семи задач тысячелетия)?','Перельман',4,6),
+    ('Какая фамилия основоположника современной гидроаэродинамики в России/СССР?','Жуковский',4,7),
+    ('Ему принадлежит серия глобальных открытий в области ядерной физики. Руководил возведением первой в мире атомной электростанции. (Фамилия)',' Курчатов',4,8),
+    ('Создатель ракетно-космической техники и практической космонавтики. Руководил  запуском первого искусственного спутника Земли и полетом первого космонавта планеты. (Фамилия)','Королев',4,9),
+    ('Офтальмолог, микрохирург. Первым в мире сделал операцию по лечению глаукомы на ранних стадиях. (Фамилия)','Федоров',4,10);
+
+INSERT INTO questions(text,answer,id_topic,id_difficulty)
+VALUES
+    ('Что означало слово «деньнощие» на Руси?','Вечер',5,1),
+    ('На каком материке находится сельва?','Южная Америка',5,2),
+    ('Какая сказка Андерсена так начинается: «Хорошо было за городом!»?','Гадкий утенок',5,3),
+    ('В описании изготовления чего в романе «Таинственный остров» Жюль Верн специально допустил ошибку?','Нитроглицерин',5,4),
+    ('На каком озере есть остров Валаам?','Ладожское',5,5),
+    ('Кто из военачальников известен как «белый генерал» у русских и «ак—паша» у турок? (Фамидия)','Скобелев',5,6),
+    ('Как на самом деле звали героиню повести «Ася» Тургенева?','Анастасия',5,7),
+    ('Где зародилась культура Возрождения?','Италия',5,8),
+    ('Где находится гробница пророка Мухаммеда?','Медина',5,9),
+    ('Как еще называют кадык на шее мужчины?','Адамово яблоко',5,10);
+
+INSERT INTO questions(text,answer,id_topic,id_difficulty)
+VALUES
+    ('Какое изобразительно-выразительное средство используется для обрисовки, пояснения, характеристики какого-либо свойства или признака предмета?','Эпитет',6,1),
+    ('Какое изобразительно-выразительное средство языка использовал В.В.Маяковский в стихотворении "Блэк энд уайт"? "Белый ест ананас спелый, черный - гнилью моченый. Белую работу делает белый, чёрную работу - чёрный."','Антитеза',6,2),
+    ('Какой вид тропа чаще всего используется в баснях?','Аллегория',6,3),
+    ('Художественный приём, основанный на повторении однородных гласных звуков, использованный Н.А.Некрасовым в стихах, откуда взяты эти строки: Кажется, шепчут колосья друг другу: "Скучно нам слушать осеннюю вьюгу."','Ассонанс',6,4),
+    ('Определите художественно-изобразительное средство. «И над отечеством свободы просвещенной. Взойдет ли наконец прекрасная заря?»','Символ',6,5),
+    ('Определите художественно-изобразительное средство. "В один прекрасный вечер, не менее прекрасный экзекутор..."','Ирония',6,6),
+    ('Какой художественный приём использовал С.Есенин в данном отрывке? "Полевая Россия! Довольно Волочиться сохой по полям! Нищету твою видеть больно И берёзам, и тополям.','Олицетворение',6,7),
+    ('Какое модернистское течение в своем манифесте «Пощечина общественному вкусу» ошеломило читателей с первого пункта: "Бросить Пушкина, Достоевского, Толстого и прочих, прочих с Парохода современности..."?','Футуризм',6,8),
+    ('Что обозначает данное определение? "Употребление слова в переносном смысле и перенесение действий и признаков одних предметов на другие, сходные с ними"','Метафора',6,9),
+    ('Определите данный литературоведческий термин. Предельное преувеличение, основанное на фантастике, на причудливом сочетании фантастического и реального.','Гротеск',6,10);
